CLAUDE.md違反を修正: 可視性・ファイル分割・unsafe除去・unwrap除去
- ExplorerPageの全pubフィールドをpub(super)に変更（可視性ルール準拠）
- explorer/mod.rs (660行→366行): event.rs, preview_open.rsに分割
- indexer.rs (420行→320行): SearchBackend implをindexer_search.rsに分割
- row.rs (267行→205行): スニペット描画をrow/snippets.rsに分離
- header.rs: render()を30行に縮小、nav/actions/breadcrumbを別関数化
- file_list.rs: unsafeブロックを除去、Rc<dyn Fn>でコールバック共有
- syntax.rs: unwrap()をlet-else早期リターンに置換
- listing.rs: spawn_blocking結果のunwrap()をResult伝播に置換
- 未使用import/関数の整理
