+++ src/pages/explorer/entries.rs
@@ -1,2 +1,2 @@
 // 純粋ロジックは core から再エクスポート
-pub use crate::core::sort::{get_extension, sort_entries};
+pub use crate::core::sort::sort_entries;
